added modularized structure
